brousla-cloud/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg[binary]==3.1.17
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
stripe==7.11.0
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
python-multipart==0.0.6
